# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Safe Place API",
    "version": "1.0.0"
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/notifications": {
      "get": {
        "summary": "Get all notifications",
        "operationId": "getAllNotifications",
        "tags": ["Notifications"],
        "responses": {
          "200": {
            "description": "A list of notifications",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Notification"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new notification",
        "operationId": "createNotification",
        "tags": ["Notifications"],
        "parameters": [
          {
            "in": "body",
            "name": "notification",
            "description": "Notification object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewNotification"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Notification created successfully",
            "schema": {
              "$ref": "#/definitions/Notification"
            }
          }
        }
      }
    },
    "/notifications/{id}": {
      "get": {
        "summary": "Get notification by ID",
        "operationId": "getNotificationById",
        "tags": ["Notifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification details",
            "schema": {
              "$ref": "#/definitions/Notification"
            }
          },
          "404": {
            "description": "Notification not found"
          }
        }
      },
      "put": {
        "summary": "Update notification by ID",
        "operationId": "updateNotification",
        "tags": ["Notifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "notification",
            "description": "Updated notification object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewNotification"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification updated successfully",
            "schema": {
              "$ref": "#/definitions/Notification"
            }
          },
          "404": {
            "description": "Notification not found"
          }
        }
      },
      "delete": {
        "summary": "Delete notification by ID",
        "operationId": "deleteNotification",
        "tags": ["Notifications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Notification deleted successfully"
          },
          "404": {
            "description": "Notification not found"
          }
        }
      }
    },
...
    "/reports": {
      "get": {
        "summary": "Get all reports",
        "operationId": "getAllReports",
        "tags": ["Reports"],
        "responses": {
          "200": {
            "description": "A list of reports",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Report"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new report",
        "operationId": "createReport",
        "tags": ["Reports"],
        "parameters": [
          {
            "in": "body",
            "name": "report",
            "description": "Report object to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewReport"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Report created successfully",
            "schema": {
              "$ref": "#/definitions/Report"
            }
          }
        }
      }
    },
    "/reports/{id}": {
      "get": {
        "summary": "Get report by ID",
        "operationId": "getReportById",
        "tags": ["Reports"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Report details",
            "schema": {
              "$ref": "#/definitions/Report"
            }
          },
          "404": {
            "description": "Report not found"
          }
        }
      },
      "put": {
        "summary": "Update report by ID",
        "operationId": "updateReport",
        "tags": ["Reports"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "report",
            "description": "Updated report object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewReport"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report updated successfully",
            "schema": {
              "$ref": "#/definitions/Report"
            }
          },
          "404": {
            "description": "Report not found"
          }
        }
      },
      "delete": {
        "summary": "Delete report by ID",
        "operationId": "deleteReport",
        "tags": ["Reports"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Report deleted successfully"
          },
          "404": {
            "description": "Report not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Notification": {
      "type": "object",
      "properties": {
        "Id_Notifications": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date_upload": {
          "type": "string"
        },
        "Id_User": {
          "type": "string"
        }
      },
      "required": ["type", "Id_User"]
    },
    "NewNotification": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "date_upload": {
          "type": "string"
        },
        "Id_User": {
          "type": "string"
        }
      },
      "required": ["type", "Id_User"]
    },
...
    "Report": {
      "type": "object",
      "properties": {
        "Id_Report": {
          "type": "string"
        },
        "Id_Reporter": {
          "type": "string"
        },
        "Id_Reported": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        },
        "Reports": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Reports"
          }
        }
      },
      "required": ["Id_Reporter", "Id_Reported"]
    },
    "NewReport": {
      "type": "object",
      "properties": {
        "Id_Reporter": {
          "type": "string"
        },
        "Id_Reported": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "reason": {
          "type": "string"
        }
      }
    },
...
  }
}